// src/app/models/condition-new-competition.ts
export interface ConditionNewCompetition {
    Successful: boolean;
    NotOwner: boolean;
    InvalidName: boolean;
    InvalidStartDate: boolean;
    InvalidEndDate: boolean;
    InvalidRangeDate: boolean;
    InvalidBid: boolean;
    InvalidPoolLength: boolean;
    InvalidPoolLanes: boolean;
    InvalidContributuon: boolean;
    InvalidMaxMembers: boolean;
    InvalidMaxComands: boolean;
    InvalidMaxComandsMembers: boolean;
    NotPay: boolean;
}